Add driver for simplejtag level 0 protocol (cross-adapter).
SimpleJTAG is a protocol designed to be easy implementable on any hardware,
targeting first of all low-cost and open-hardware development boards.
SimpleJTAG is expected to provide several levels of protocol, offering
different ratio of ease of firmware development vs target access speed.
This driver is for "level 0" protocol, which uses the simplest possible
bitbanging protol, where each bit transition is driven by host. It is also
half-duplex, following request-response pattern for each command. So, it
is slow, but at the same time, can be implemented even on the simplest
MCUs, with minimum flash space, and with bitbang UART implementation for
connection to host.

Cross-MCU and cross-board implementation of the protocol can be found at
https://github.com/pfalcon/simplejtag

Signed-off-by: Paul Sokolovsky <pfalcon@users.sourceforge.net>